chore: add typeguard and complexities
